+++ CHANGES.rst
@@ -7,6 +7,7 @@ Development
 * Add get_current_metrics to get the metrics utility from the thread-local
   data if the current request object is not otherwise available (for example,
   in the context of a model method.)
+* Raise an exception if get_current_metrics is called and no request is active
 
 0.1.4 (2014-05-08)
 ------------------